fix(pty/sshfs): split linux-only helpers to identity_linux.go (unused lint)
darwin/windows golangci-lint flagged sshfsDestRE / parseSSHFSDestination /
injectDefaultPort / resolveSSHFSIdentity as unused — they only have callers
in mount_linux.go, which isn't built on those platforms.

Move them to a new identity_linux.go with //go:build linux. The flag vars
(sshfsSK / sshfsNoVisor / sshfsDefPort) stay in sshfs.go because init() on
all platforms registers them on RootCmd.

E2E test failure on linux (TestMux/mux_distributes_traffic_across_STCPR_direct_+_via-b
in internal/integration) is unrelated to this PR — pre-existing in develop.
